# Inline intermediate operations using `yield`