--- a/src/main/php/util/data/FilterableWithKey.class.php
+++ b/src/main/php/util/data/FilterableWithKey.class.php
@@ -4,6 +4,8 @@
  * A filterable only returns elements which the given accept function
  * returns true for, omitting the others.
  *
+ * @deprecated
+ * @test  xp://util.data.unittest.FlattenerTest
  * @see   php://CallbackFilterIterator but only passes two arguments
  */
 class FilterableWithKey extends AbstractFilterable {