--- a/src/main/php/util/data/Mapper.class.php
+++ b/src/main/php/util/data/Mapper.class.php
@@ -4,6 +4,7 @@
  * A mapper reaches applies a given mapper function to each value an
  * iterator returns and returns its result.
  *
+ * @deprecated
  * @test  xp://util.data.unittest.MapperTest
  */
 class Mapper extends AbstractMapper {