--- a/src/main/php/util/data/Window.class.php
+++ b/src/main/php/util/data/Window.class.php
@@ -5,6 +5,7 @@
  * iterator that match its "skip" closure, then returns elements
  * until its "stop" closure is reached.
  *
+ * @deprecated
  * @see   php://LimitIterator but uses closures and not just offsets
  * @test  xp://util.data.unittest.WindowTest
  */