--- a/src/test/php/util/data/unittest/FilterableTest.class.php
+++ b/src/test/php/util/data/unittest/FilterableTest.class.php
@@ -2,6 +2,7 @@
 
 use util\data\Filterable;
 
+/** @deprecated */
 class FilterableTest extends \unittest\TestCase {
 
   /**