--- a/src/test/php/util/data/unittest/FlattenerTest.class.php
+++ b/src/test/php/util/data/unittest/FlattenerTest.class.php
@@ -3,6 +3,7 @@
 use util\data\Flattener;
 use util\data\Sequence;
 
+/** @deprecated */
 class FlattenerTest extends \unittest\TestCase {
 
   /**