--- a/src/test/php/util/data/unittest/MapperTest.class.php
+++ b/src/test/php/util/data/unittest/MapperTest.class.php
@@ -3,6 +3,7 @@
 use util\data\Mapper;
 use lang\IllegalStateException;
 
+/** @deprecated */
 class MapperTest extends \unittest\TestCase {
 
   /**