## Performance comparison

Uses the class from [this gist](https://gist.github.com/thekid/825bbb7bf142476d614022176755ee1d)

```bash
# Before
$ xp -m ..\measure util.profiling.Measure Operations -n 1000
filter: 1000 iteration(s), 5.998 seconds, result= 5000
map: 1000 iteration(s), 7.635 seconds, result= 10000
peek: 1000 iteration(s), 8.342 seconds, result= 10000
flatten: 1000 iteration(s), 27.544 seconds, result= 20000
skip_n: 1000 iteration(s), 1.934 seconds, result= 9000
skip: 1000 iteration(s), 7.600 seconds, result= 9000
limit_n: 1000 iteration(s), 1.874 seconds, result= 9000
limit: 1000 iteration(s), 6.867 seconds, result= 9000

# After
$ xp -m ..\measure util.profiling.Measure Operations -n 1000
filter: 1000 iteration(s), 1.443 seconds, result= 5000
map: 1000 iteration(s), 1.558 seconds, result= 10000
peek: 1000 iteration(s), 1.317 seconds, result= 10000
flatten: 1000 iteration(s), 3.571 seconds, result= 20000
skip_n: 1000 iteration(s), 0.928 seconds, result= 9000
skip: 1000 iteration(s), 0.872 seconds, result= 9000
limit_n: 1000 iteration(s), 0.898 seconds, result= 9000
limit: 1000 iteration(s), 1.323 seconds, result= 9000
```

## Functions and their speedup

* [x] filter() - factor 4.1
* [x] map() - factor 4.9
* [x] peek() - factor 6.3
* [x] flatten() - factor 7.7
* [x] skip() - factor 2 (integer), factor 8.7 (function)
* [x] limit() - factor 2 (integer), factor 5.1 (function)