@thekid 2017-04-25
Files which could be removed (*or at least deprecated*):

* src/main/php/util/data/AbstractWindow.class.php
* src/main/php/util/data/Filterable.class.php
* src/main/php/util/data/FilterableWithKey.class.php
* src/main/php/util/data/Flattener.class.php
* src/main/php/util/data/Mapper.class.php
* src/main/php/util/data/MapperWithKey.class.php
* src/main/php/util/data/Window.class.php
* src/main/php/util/data/WindowWithKey.class.php
* src/test/php/util/data/unittest/FilterableTest.class.php
* src/test/php/util/data/unittest/FlattenerTest.class.php
* src/test/php/util/data/unittest/MapperTest.class.php
* src/test/php/util/data/unittest/WindowTest.class.php